@@ -2,17 +2,19 @@
 
 namespace Stringy;
 
-if (!function_exists('Stringy\create')) {
+if (!function_exists(__NAMESPACE__ . '\create')) {
     /**
      * Creates a Stringy object and returns it on success.
      *
-     * @param  mixed   $str      Value to modify, after being cast to string
-     * @param  string  $encoding The character encoding
-     * @return Stringy A Stringy object
-     * @throws \InvalidArgumentException if an array or object without a
-     *         __toString method is passed as the first argument
+     * @param string|\Stringable $str
+     *   Value to modify, after being cast to string.
+     * @param string $encoding
+     *   The character encoding.
+     *
+     * @return \Stringy\Stringy
+     *   A Stringy instance.
      */
-    function create($str, $encoding = null)
+    function create(string|\Stringable $str, string $encoding = null)
     {
         return new Stringy($str, $encoding);
     }
